+++ spec/fabricators/sample.rb
@@ -1,0 +1,7 @@
+Fabricator(:sample, from: Sample, aliases: [:valid_sample]) do
+  name 'Sample'
+end
+
+Fabricator(:sample2, from: Sample) do
+  name 'Sample2'
+end