+++ spec/support/models/dummy.rb
@@ -1,9 +1,0 @@
-# = Dummy
-#
-# Dummy class for testing purpose.
-class Dummy
-  include Mongoid::Document
-  include Mongoid::UUID
-
-  field :name
-end